requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
-e git+https://github.com/ashishki/Dream_Motif_Interpreter@de77439301471d9c8e6cd0a94a62b8312a4f48c4#egg=dream_motif_interpreter
exceptiongroup==1.3.1
fastapi==0.135.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
opentelemetry-api==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pytest==8.4.2
pytest-asyncio==0.23.8
python-dotenv==1.2.2
PyYAML==6.0.3
ruff==0.15.10
starlette==1.0.0
structlog==25.5.0
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
zipp==3.23.0
